pyproject.toml
[tool.poetry]
name = "test_task"
version = "1.0.0"
authors = ["mikhailok"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
